+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+
+* Updated file storage interface.
+* Handling exceptions of caching images to the file storage.
+
 ## 0.0.1
 
 * Initial release.